requirements.txt
discord.py
discord-ext-voice-recv
pynacl